Move implementation of Assertions into separate classes
