+++ junit-jupiter-api/src/main/java/org/junit/jupiter/api/AssertAll.java
@@ -1,0 +1,58 @@
+/*
+ * Copyright 2015-2016 the original author or authors.
+ *
+ * All rights reserved. This program and the accompanying materials are
+ * made available under the terms of the Eclipse Public License v1.0 which
+ * accompanies this distribution and is available at
+ *
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+
+package org.junit.jupiter.api;
+
+import java.util.Arrays;
+import java.util.stream.Stream;
+
+import org.junit.jupiter.api.function.Executable;
+import org.junit.platform.commons.util.ExceptionUtils;
+import org.junit.platform.commons.util.Preconditions;
+import org.opentest4j.MultipleFailuresError;
+
+public class AssertAll {
+
+	public static void assertAll(Executable... executables) {
+		assertAll(null, executables);
+	}
+
+	public static void assertAll(String heading, Executable... executables) {
+		Preconditions.notEmpty(executables, "executables array must not be null or empty");
+		Preconditions.containsNoNullElements(executables, "individual executables must not be null");
+		assertAll(heading, Arrays.stream(executables));
+	}
+
+	public static void assertAll(Stream<Executable> executables) {
+		assertAll(null, executables);
+	}
+
+	public static void assertAll(String heading, Stream<Executable> executables) {
+		Preconditions.notNull(executables, "executables must not be null");
+		MultipleFailuresError multipleFailuresError = new MultipleFailuresError(heading);
+
+		executables.forEach(executable -> {
+			try {
+				executable.execute();
+			}
+			catch (AssertionError assertionError) {
+				multipleFailuresError.addFailure(assertionError);
+			}
+			catch (Throwable t) {
+				ExceptionUtils.throwAsUncheckedException(t);
+			}
+		});
+
+		if (multipleFailuresError.hasFailures()) {
+			throw multipleFailuresError;
+		}
+	}
+
+}
