+++ junit-jupiter-api/src/main/java/org/junit/jupiter/api/AssertFalse.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2015-2016 the original author or authors.
+ *
+ * All rights reserved. This program and the accompanying materials are
+ * made available under the terms of the Eclipse Public License v1.0 which
+ * accompanies this distribution and is available at
+ *
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+
+package org.junit.jupiter.api;
+
+import static org.junit.jupiter.api.Fail.fail;
+
+import java.util.function.BooleanSupplier;
+import java.util.function.Supplier;
+
+public class AssertFalse {
+
+	public static void assertFalse(boolean condition) {
+		assertFalse(() -> condition, () -> null);
+	}
+
+	public static void assertFalse(boolean condition, String message) {
+		assertFalse(() -> condition, () -> message);
+	}
+
+	public static void assertFalse(boolean condition, Supplier<String> messageSupplier) {
+		assertFalse(() -> condition, messageSupplier);
+	}
+
+	public static void assertFalse(BooleanSupplier booleanSupplier) {
+		assertFalse(booleanSupplier, () -> null);
+	}
+
+	public static void assertFalse(BooleanSupplier booleanSupplier, String message) {
+		assertFalse(booleanSupplier, () -> message);
+	}
+
+	public static void assertFalse(BooleanSupplier booleanSupplier, Supplier<String> messageSupplier) {
+		if (booleanSupplier.getAsBoolean()) {
+			fail(messageSupplier);
+		}
+	}
+
+}
