+++ junit-jupiter-api/src/main/java/org/junit/jupiter/api/AssertNotEquals.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2015-2016 the original author or authors.
+ *
+ * All rights reserved. This program and the accompanying materials are
+ * made available under the terms of the Eclipse Public License v1.0 which
+ * accompanies this distribution and is available at
+ *
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+
+package org.junit.jupiter.api;
+
+import static org.junit.jupiter.api.AssertionsUtil.buildPrefix;
+import static org.junit.jupiter.api.AssertionsUtil.nullSafeGet;
+import static org.junit.jupiter.api.AssertionsUtil.objectsAreEqual;
+import static org.junit.jupiter.api.Fail.fail;
+
+import java.util.function.Supplier;
+
+public class AssertNotEquals {
+
+	public static void assertNotEquals(Object unexpected, Object actual) {
+		assertNotEquals(unexpected, actual, () -> null);
+	}
+
+	public static void assertNotEquals(Object unexpected, Object actual, String message) {
+		assertNotEquals(unexpected, actual, () -> message);
+	}
+
+	public static void assertNotEquals(Object unexpected, Object actual, Supplier<String> messageSupplier) {
+		if (objectsAreEqual(unexpected, actual)) {
+			failEqual(actual, nullSafeGet(messageSupplier));
+		}
+	}
+
+	private static void failEqual(Object actual, String message) {
+		fail(buildPrefix(message) + "expected: not equal but was: <" + actual + ">");
+	}
+}
