+++ junit-jupiter-api/src/main/java/org/junit/jupiter/api/AssertNotSame.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2015-2016 the original author or authors.
+ *
+ * All rights reserved. This program and the accompanying materials are
+ * made available under the terms of the Eclipse Public License v1.0 which
+ * accompanies this distribution and is available at
+ *
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+
+package org.junit.jupiter.api;
+
+import static org.junit.jupiter.api.AssertionsUtil.buildPrefix;
+import static org.junit.jupiter.api.AssertionsUtil.nullSafeGet;
+import static org.junit.jupiter.api.Fail.fail;
+
+import java.util.function.Supplier;
+
+public class AssertNotSame {
+
+	public static void assertNotSame(Object unexpected, Object actual) {
+		assertNotSame(unexpected, actual, () -> null);
+	}
+
+	public static void assertNotSame(Object unexpected, Object actual, String message) {
+		assertNotSame(unexpected, actual, () -> message);
+	}
+
+	public static void assertNotSame(Object unexpected, Object actual, Supplier<String> messageSupplier) {
+		if (unexpected == actual) {
+			failSame(actual, nullSafeGet(messageSupplier));
+		}
+	}
+
+	private static void failSame(Object actual, String message) {
+		fail(buildPrefix(message) + "expected: not same but was: <" + actual + ">");
+	}
+
+}
