+++ junit-jupiter-api/src/main/java/org/junit/jupiter/api/AssertThrows.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2015-2016 the original author or authors.
+ *
+ * All rights reserved. This program and the accompanying materials are
+ * made available under the terms of the Eclipse Public License v1.0 which
+ * accompanies this distribution and is available at
+ *
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+
+package org.junit.jupiter.api;
+
+import static org.junit.jupiter.api.AssertionsUtil.format;
+
+import org.junit.jupiter.api.function.Executable;
+import org.opentest4j.AssertionFailedError;
+
+public class AssertThrows {
+
+	public static <T extends Throwable> T expectThrows(Class<T> expectedType, Executable executable) {
+		return assertThrows(expectedType, executable);
+	}
+
+	@SuppressWarnings("unchecked")
+	public static <T extends Throwable> T assertThrows(Class<? extends Throwable> expectedType, Executable executable) {
+		try {
+			executable.execute();
+		}
+		catch (Throwable actualException) {
+			if (expectedType.isInstance(actualException)) {
+				return (T) actualException;
+			}
+			else {
+				String message = format(expectedType.getName(), actualException.getClass().getName(),
+					"Unexpected exception type thrown");
+				throw new AssertionFailedError(message, actualException);
+			}
+		}
+		throw new AssertionFailedError(
+			String.format("Expected %s to be thrown, but nothing was thrown.", expectedType.getName()));
+	}
+
+}
