+++ junit-jupiter-api/src/main/java/org/junit/jupiter/api/AssertTrue.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2015-2016 the original author or authors.
+ *
+ * All rights reserved. This program and the accompanying materials are
+ * made available under the terms of the Eclipse Public License v1.0 which
+ * accompanies this distribution and is available at
+ *
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+
+package org.junit.jupiter.api;
+
+import static org.junit.jupiter.api.Fail.fail;
+
+import java.util.function.BooleanSupplier;
+import java.util.function.Supplier;
+
+public class AssertTrue {
+
+	public static void assertTrue(boolean condition) {
+		assertTrue(() -> condition, () -> null);
+	}
+
+	public static void assertTrue(boolean condition, Supplier<String> messageSupplier) {
+		assertTrue(() -> condition, messageSupplier);
+	}
+
+	public static void assertTrue(BooleanSupplier booleanSupplier) {
+		assertTrue(booleanSupplier, () -> null);
+	}
+
+	public static void assertTrue(BooleanSupplier booleanSupplier, String message) {
+		assertTrue(booleanSupplier, () -> message);
+	}
+
+	public static void assertTrue(boolean condition, String message) {
+		assertTrue(() -> condition, () -> message);
+	}
+
+	public static void assertTrue(BooleanSupplier booleanSupplier, Supplier<String> messageSupplier) {
+		if (!booleanSupplier.getAsBoolean()) {
+			fail(messageSupplier);
+		}
+	}
+
+}
